+++ code/qcommon/vm_x86.c
@@ -1423,21 +1423,22 @@ emit_comiss(uint32_t base, uint32_t reg)
 }
 
 static void
-emit_load_sx(uint32_t reg, uint32_t base, int32_t offset)
+emit_load_sx(uint32_t xmmreg, uint32_t base, int32_t offset)
 {
   Emit1(0xF3);
-  emit_op_reg_base_offset(0x0F, 0x10, reg, base, offset);
+  emit_op_reg_base_offset(0x0F, 0x10, xmmreg, base, offset);
 }
 
 static void
-emit_load_sx_index(uint32_t reg, uint32_t base, uint32_t index)
+emit_load_sx_index(uint32_t xmmreg, uint32_t base, uint32_t index)
 {
   Emit1(0xF3);
-  emit_op_reg_base_index(0x0F, 0x10, reg, base, index, 1, 0);
+  emit_op_reg_base_index(0x0F, 0x10, xmmreg, base, index, 1, 0);
 }
 
+//wrapper function
 static void
-emit_mov_sx_rx(uint32_t xmmreg, uint32_t intreg)
+mov_sx_rx(uint32_t xmmreg, uint32_t intreg)
 {
   if (CPU_Flags & CPU_SSE2)
   {
@@ -1448,8 +1449,8 @@ emit_mov_sx_rx(uint32_t xmmreg, uint32_t intreg)
   else
   {
     //SSE1 CPUs do not initialize FP register domain with movd so use movss
-    emit_store_rx(intreg, R_PROCBASE, 0); //mov [procBase + 0], eax
-    emit_load_sx(xmmreg, R_PROCBASE, 0); //mov xmm0, [procBase + 0]
+    emit_store_rx(intreg, R_PROCBASE, 0); //mov dword ptr [procBase + 0], eax
+    emit_load_sx(xmmreg, R_PROCBASE, 0); //movss xmm0, [procBase + 0]
   }
 }
 
@@ -1497,6 +1498,7 @@ emit_mul_sx(uint32_t dst, uint32_t src)
 static void
 emit_div_sx(uint32_t dst, uint32_t src)
 {
+  Emit1(0xF3); //use divss instead of divps to avoid division by zero from src[32...127] and triggering "invalid operation" bit
   emit_op_reg(0x0F, 0x5E, src, dst);
 }
 
@@ -1642,13 +1644,13 @@ alloc_sx(uint32_t pref);
 
 //register allocation preferences
 
-#define FORCED 0x20  //load function must return specified register
+#define FORCED 0x20  //load function must return specified register, no dynamic allocation allowed
 #define TEMP 0x40  //hint: temporary allocation, will not be stored on opStack
 #define RCONST 0x80  //register value will be not modified
 #define XMASK 0x100 //exclude masked registers
 #define SHIFT4 0x200 //load bottom item
 
-#define RMASK 0x0F
+#define RMASK 0x0F //mask to get register number from preference
 
 //array sizes for cached/meta registers
 #if idx64
@@ -2159,7 +2161,7 @@ mov_sx_imm(uint32_t reg, uint32_t imm32)
   else
   {
     uint32_t rx = alloc_rx_const(R_ECX | TEMP, imm32); //ecx = imm32
-    emit_mov_sx_rx(reg, rx); //xmmX = ecx
+    mov_sx_rx(reg, rx); //xmmX = ecx
     unmask_rx(rx);
   }
 }
@@ -3405,7 +3407,7 @@ load_sx_opstack(uint32_t pref)
     //should never happen with FPU type promotion, except syscalls
     reg = alloc_sx(pref);
 
-    emit_mov_sx_rx(reg, it->value);
+    mov_sx_rx(reg, it->value);
 
     it->type = TYPE_RAW;
     return reg;
@@ -3432,7 +3434,7 @@ load_sx_opstack(uint32_t pref)
     reg = alloc_sx(pref);
     rx = alloc_rx_local(R_ECX | RCONST, it->value);
 
-    emit_mov_sx_rx(reg, rx); //move from integer to scalar
+    mov_sx_rx(reg, rx); //move from integer to scalar
 
     unmask_rx(rx);